01-python-mcp-server/auth.py
# auth.py
import os
from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow

SCOPES = ["https://www.googleapis.com/auth/calendar"]  # Full access for creating events


def get_credentials():
    """
    Authenticate with Google Calendar API and return valid credentials.
    Saves token.json for future use.
    """
    creds = None

    if os.path.exists("token.json"):
        creds = Credentials.from_authorized_user_file("token.json", SCOPES)

    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file("credentials.json", SCOPES)
            creds = flow.run_local_server(port=0)

        # Save the credentials for next run
        with open("token.json", "w") as token_file:
            token_file.write(creds.to_json())

    return creds


if __name__ == "__main__":
    creds = get_credentials()
    print("✅ Authentication successful. Token saved to token.json")
